Code Review Comments
I have to say, I really love the story your page captures (and I really mean that), it is so unique and I can see what you were going for. The content was great and not many people have such a unique background, I know this is just an initial project for this class, but you definitely have to build it further and fully develop that theme of "From pianist to programmer". I added a couple comments to the CSS file, the first thing is that it doesn't hurt to break it up into different files for different pages, it will keep it more organized especially as the project scales. The other thing is that you can format the whole page with prettier to apply the consistent spacing and formatting. Really cool webpage and your background is so unique so building on that is a definite strong point! Very memorable :)